Update to handle translated interface.
These are translations of the user interface for PiCom.

They have been machine translated and would probably benefit from a native speaker editing them.

As they are machine translated we do not expect them to be perfect in any way, but hopefully they will provide a starting point for some users.